Allow objects.CreateTempURL with names containing '/v1/'
Before this patch, if an object's name or container name contained the
string `/v1/`, it was not possible to create a TempURL for it.
This was due to a bug in how the URL was split between the base part,
and the object-path part.

With this patch, container names or object names containing `/v1/` can
have a TempURL created successfully.