Fix networking acceptance tests
Since https://review.opendev.org/c/openstack/neutron/+/892815 it is no
longer possible to update the vnic type of a bound port.

These doesn't seem to be an API to unbind the port other than deleting
the port, so instead we're dropping the `VNICType` attribute from the
update request in the `TestPortsbindingCRUD` test.