Fix value_specs implementation
The current implementation does not do anything useful. The key-pairs
must be extracted and added directly in the body instead of under
"value_specs".
This commit fixes the implementation and also adds validation of the
value_specs based on what is done in Heat. There are some banned keys
that are not allowed in value_specs, and values are not allowed to
overwrite existing fields either.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>