remove the error types for specific response codes
This is a breaking change intended for the v2.0 release. Rationale:

The ErrDefault series of types (ErrDefault400, ErrDefault401, etc.)
serve the apparent purpose of allowing to assert on a specific response
code (e.g. to handle 404 responses differently from other 4xx/5xx
responses). However, because of their design, they often tend to obscure
the real underlying error message in many cases, reporting just
something like "Method not allowed" instead of the default error string
on ErrUnexpectedResponseCode that explains which request failed and
includes the actual error message returned by the OpenStack service.

This commit removes the types, and all of the associated boilerplate,
and replaces them with a single small helper function for quickly
asserting that an error is an ErrUnexpectedResponseCode with a specific
response code.

This also gets rid of the undocumented and untested
ProviderClient.ErrorContext interface that I have legitimately no idea
what it's supposed to be good for, besides maybe skipping the generation
of ErrDefaultXXX-type errors.