Fix AllowReauth reauthentication
Due to an error in implementing the addition of `context.Context`, the
reauth function catched the context passed when generating the
ProviderClient, which could be long canceled when the reauthentication
takes place.

This is a breaking change: this patch changes the signature of the
reauthentication function in the ProviderClient to accept a context.