blockstorage: Merge standalone extensions
Extensions are no longer a thing in any OpenStack service except
Neutron. As such, maintaining what were formerly extensions in a
separate module makes little sense and actively hurts us as these
extensions are themselves extended with new microversions.

Resolve this for the block storage service first. In this case, we
simply migrate the various standalone extensions (i.e. the extensions
that did not build on existing resources) into the v2 and v3 modules.
This appears like we are creating a lot of new code, but we simply
ran the following Python script (because Bash arrays are awful):

    import shutil
    for extension in {
        'availabilityzones',
        'backups',
        'limits',
        'quotasets',
        'schedulerstats',
        'services',
    }:
        src = f'openstack/blockstorage/extensions/{extension}'
        for api_version in {'v2', 'v3'}:
            dst = f'openstack/blockstorage/{api_version}/{extension}'
            shutil.copytree(src, dst)
        shutil.rmtree(src)

We then fix the imports in the new files and then 'git add .'

Once this is done, we need to do something very similar for the tests
for these "extensions" in 'internal/acceptance/blockstorage/extensions'.
Once again we copy-paste the code to both v2 and v3 modules, but this
time we must both update imports and modify the v2 variants of the tests
to use a v2 client.

Only volumetransfers is excluded since we want to rename that module.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>