Add minimal Makefile
To quickly run linters, unit tests and potentially acceptance tests
against a local DevStack. We will eventually add golangci-lint to the
'lint' target but that's a larger piece of work that should be done
separately. We should also rework the GitHub Actions jobs to use these
targets and remove the scripts, but that requires more work on job skips
to be able to do that.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>