compute: Merge standalone extensions
Extensions are no longer a thing in any OpenStack service except
Neutron. As such, maintaining what were formerly extensions in a
separate module makes little sense and actively hurts us as these
extensions are themselves extended with new microversions.

Resolve this for the compute service now. In this case, we
simply migrate the various standalone extensions (i.e. the extensions
that did not build on existing resources) into the v2 modules.
This appears like we are creating a lot of new code, but we simply
ran the following Python script (because Bash arrays are awful):

    import shutil
    for extension in {
        'aggregates',
        'attachinterfaces',
        'defsecrules',
        'diagnostics',
        'floatingips',
        'hypervisors',
        'instanceactions',
        'limits',
        'networks',
        'quotasets',
        'remoteconsoles',
        'secgroups',
        'servergroups',
        'services',
        'tags',
        'tenantnetworks',
        'usage',
        'volumeattach',
    }:
        src = f'openstack/compute/v2/extensions/{extension}'
        dst = f'openstack/compute/v2/{extension}'
        shutil.move(src, dst)

We then fix the imports in the new files and then 'git add .'

Once this is done, we need to do something very similar for the tests
for these "extensions" in 'internal/acceptance/compute/extensions'.
Once again we move the code to the v2 modules, but this
time we must update imports.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>