identity: Merge standalone v3 extensions
Extensions are no longer a thing in any OpenStack service except
Neutron. As such, maintaining what were formerly extensions in a
separate module makes little sense and actively hurts us as these
extensions are themselves extended with new microversions.

It's the identity services turn now and we do as we've done for
blockstorage and compute previously, moving these extensions up a level
to the root 'openstack/identity/v3' module, updating references in the
process.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>