Drop CI coverage for Victoria-Yoga
Yoga has entered, and the rest will soon enter, the unmaintained [1]
phase. This means that the CI stability will be in the hands of a small
group of volunteers and will no longer be the responsibility of the
project teams. I don't believe the Gophercloud project can maintain CI
jobs on these branches in the long run, hence removing them.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html#unmaintained