tests: Fix lint issues
We have a number of tests that we do not appear to be running in CI. In
addition, we are not running go vet with all suitable build tags
resulting in some tests being skipped [1]. Combined, this has resulted
in a number of tests that have not been fully updated to use context.
While we will address the root cause elsewhere, we can fix the missed
updates here.

You can validate this fix by overriding GOFLAGS, as outlined in [1]. For
example, from the root directory:

  GOFLAGS="-tags=acceptance" go vet ./...

[1] https://github.com/golang/go/issues/29202

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>