acceptance: Add 'RequireNeutronExtension' helper
Skip neutron tests if the relevant extension is not available in our
environment. We migrate existing variations of this to the new helper
and add some for tests that depend on neutron's 'fwaas_v2' extension
(though not for fwaas v1, since those tests are all already skipped and
should be removed in a separate change).

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>