script/stackenv: Don't set 'OS_SYSTEM_SCOPE' until necessary
We were setting OS_SYSTEM_SCOPE in some of the functional ironic tests.
This was resulting in the following errors when attempting to create
resources:

  Authentication cannot be scoped to multiple targets. Pick one of:
  project, domain, trust, system or unscoped

Unfortunately, because we weren't setting 'errexit' ('set -e') this
wasn't being noticed, but we can "correct" things now. I say "correct"
because, adding to the misfortune, how we fix this is technically
incorrect - you should have project-scoped configuration *or*
system-scoped, not both - but since we don't yet have the infrastructure
to switch between sets of auth info on a test-by-test basis, we need to
do a Bad Thing. Hopefully this will be a short-lived thing.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>