tests: Remove service update integration test
Go runs tests in parallel for different packages in parallel. The
service update test was disabling and re-enabling all compute services,
and if this ran at the same time as another test that created a server
then the server creation (and therefore the test) would fail. There does
not appear to be any way to resolve this. We could run all integration
tests serially instead, but this would be a backwards steps that will
massively increase test run time. We could also rewrite the test to only
disable 1 of N nova-compute services, but that would require us to fence
off this compute node in all other tests using non-default parameters to
server create. Given the lack of better options, we simply remove the
test and assume our testing to date on this method has been good enough
and unit tests will be enough to carry us forward.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>