Configure EndpointOpts.Aliases
This allows us to set the service types to their official value rather
than one of the aliases.

There is some level of duplication going on here, in that our aliases
are defined in two places. That's okay since it covers the many
different ways a user can create a service client.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>