lint: Remove non-constant format string in calls (govet)
We were incorrectly using 'fmt.Printf', 'fmt.Errorf' and 't.Logf' with
non-template strings/no arguments. The fix to this is replace these
calls with the non-suffixed variants. There are many users of
'fmt.Fprint' - too many to do by hand - so this replacement was resolved
using 'sed':

  sed 's/Fprintf/Fprint/g' -i $(ag fmt.Fprintf -l)

We then manually fix the 25 cases where 'fmt.Fprintf' is actually
warranted and manually replaced the errant users of 'fmt.Errorf' and
't.Logf'. We also rework 'internal/acceptance/clients/clients.go'
slightly to make the code a bit clearer.

PS: This is apparently going to be an issue in go 1.24 (specifically in
'go vet') [1] so this is not just golangci-lint being annoying.
@PierrePrinetti, that's directed at you ;)

[1] https://github.com/golang/go/issues/60529

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>