trivial: Simplify endpoint filtering
We do not need to check the validity of the provided opts more than once
so don't. We can also simplify our handling of multiple endpoints
(though not as much as we'd like in the v2 case, due to forthcoming
patches).

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 43294956db79da1ba7e02f169494ebd29a3040be)