Prefer versioned service type aliases
This is a partial revert of e3947338f69b8b52bd494f09388152febc49c6aa.
Prefer e.g. volumev3 or sharev2 to block-storage or shared-file-systems
as a temporary measure while we work on versioned API discovery. Couple
with prior changes, this means if we e.g. have two catalog entries with
services types of 'volumev2' and 'volumev3' (in that order!) and we
request a service with 'volumev3', we will return the latter catalog
entry ahead of the former.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>