Sort discovered endpoints in order of service types
Currently, we iterate through endpoints first, comparing them to the
request service type and its aliases. This means that endpoints which
use an alias can end up being selected ahead of endpoints that use the
requested type. Invert this logic so that we prefer endpoints that match
our requested type ahead of those that match an alias of the requested
type.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>