tests: Fix TestBGPAgentCRUD
Change our deployment so that we use the static scheduler, and rework
the test to handle this. Instead of waiting for the speaker to be
associated with an agent (which won't happen with the static scheduler)
we now jump straight to assigning it.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes: #3380
(cherry picked from commit 669a870ec02690a2adc7945a840e22fbf4933d62)