Migrate to tools
The tool directive is available since Go 1.24 and is the official
replacement for `go run pkg@version`. It allows us to pin the tool and
its dependencies via go.mod/go.sum and allow them to be verified.

Migrate to this directive for the test runner. We only migrate the test
runner since the golangci-lint docs are *very* clear that you shouldn't
use 'go tool' for that tool [1].

[1] https://golangci-lint.run/welcome/install/#install-from-sources

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>