Fix pagination for messaging client
Zaqar returns paths rather than URLs in 'links' response fields and the
'Location' header. In addition, since Dalmatian, it deploys under a path
('/messaging') rather then behind a port. This combination has
highlighted a bug in our pagination page handling, namely that we are
ignoring the path element of the endpoint URL. While this is not an
issue when the 'next' link is an absolute URL, as with most other
services, it results in incomplete URLs when the link URL is relative
('/v2/messages') and the endpoint URL contains a path
('https://example.com/messaging').

Fix this by concatenating the two paths, rather than overwriting the
endpoint URLs path.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>